src/requirements.txt
asgiref==3.6.0
Django==4.1
djangorestframework==3.14.0
pyotp==2.8.0
python-environ==0.4.54
pytz==2023.3
sqlparse==0.4.4
